// src/types/discotype.ts
export type vinilo = {
	id: string;
	image: string;
	album: string;
	artista: string;
	precio: string;
	stock: string;

};

export type viniloform ={	
	image: string;
	album: string;
	artista: string;
	precio: string;
	stock: string;

};

export enum Coleccion {
	'vinilos' = 'vinilos',
}